🐛 Fixed the list like bug

+++ api/services/list_service.py
@@ -311,9 +311,11 @@ def toggle_like_list(user, share_code):
     if list_obj is None:
         return False
 
-    try:
+    like = ListLike.objects.filter(user=user, list=list_obj).first()
+
+    if like:
         ListLike.objects.filter(user=user, list=list_obj).delete()
-    except ListLike.DoesNotExist:
+    else:
         like = ListLike(user=user, list=list_obj)
         Notification.create(1, NotificationTypes.NEW_LIST_LIKE.object, list_obj.owner, user.share_code)
         like.save()
